+++ frontend/src/components/common/ActionButton.tsx
@@ -2,19 +2,17 @@ const ActionButton = ({
   onClick,
   color,
   children,
-  isLoading = false,
 }: {
   onClick: () => void;
   color: string;
   children: React.ReactNode;
-  isLoading?: boolean;
 }) => {
   return (
     <button
       onClick={onClick}
-      className={`${color} text-white rounded-lg px-4 py-2 flex items-center gap-1`}
+      className={`${color} text-white rounded-lg px-4 py-2 flex items-center justify-center gap-1`}
     >
-      {isLoading ? "Loading..." : children}
+      {children}
     </button>
   );
 };
